backend/requirements.txt
# fastapi==0.110.0
# uvicorn[standard]==0.29.0

# LLM Providers
# openai==1.30.0
# google-generativeai==0.5.4      # Uncomment for Gemini
# anthropic==0.25.0               # Uncomment for Claude

# Database
# asyncpg==0.29.0
# sqlalchemy[asyncio]==2.0.29
# alembic==1.13.1

# Core
# pydantic==2.7.0
# pydantic-settings==2.2.1
# python-dotenv==1.0.1
# httpx==0.27.0

# WhatsApp / Twilio
# twilio==8.13.0

# Utils
# python-multipart==0.0.9
# slowapi==0.1.9
# tenacity==8.2.3
# structlog==24.1.0
# python-dateutil==2.9.0

# Testing
# pytest==8.1.1
# pytest-asyncio==0.23.6

fastapi==0.115.0
uvicorn[standard]==0.29.0
openai==1.30.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.29
alembic==1.13.1
pydantic==2.11.7
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
twilio==8.13.0
python-multipart==0.0.9
slowapi==0.1.9
tenacity==8.2.3
structlog==24.1.0
python-dateutil==2.9.0
pytest==8.1.1
pytest-asyncio==0.23.6
